test: only check headers for text mode

+++ run-test.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-# rm -rf build
 set -e
+laddr=127.0.0.1
 lport=1081
 sport=8888
 
@@ -21,7 +21,7 @@ check() {
     rm -rf o.*
     echo "checking donwload binary"
     curl -q -m 6 -4 -s -L $urlb -o o.bin
-    curl -q -m 6 -4 -s -x socks5h://localhost:$lport -L $urlb -o o.ws.bin
+    curl -q -m 6 -4 -s -x socks5h://$laddr:$lport -L $urlb -o o.ws.bin
     if ! cmp -s o.bin o.ws.bin; then
         echo "fail test binary"
         exit 1
@@ -29,21 +29,23 @@ check() {
     echo "download binary ok"
 
     echo "checking upload text"
-    header=`head -c 4096 /dev/urandom | openssl base64`
-    magic=`echo $header | openssl base64 -d | openssl md5 | awk '{print $NF}'`
-    curl -q -m 6 -4 -s -L -H "X-magic: $magic" $urlt?mask={1,12,123,1234} https://$urlt?mask={1,12,123,1234} -o - > o.magic.txt
+    head -c 4096 /dev/urandom | openssl base64 | cat -n | sed -e 's|^ *|X-header-|g' -e 's|\t|: |g' > o.headers.txt
+    magic=`head -c 42 /dev/urandom | openssl md5 | awk '{print $NF}'`
+    curl -q -m 6 -4 -s -L -H "X-magic: $magic" $urlt?mask={1,12,123,1234} https://$urlt?mask={1,12,123,1234} -o - | grep '^X-magic:' > o.magic.txt
     if ! grep $magic o.magic.txt >/dev/null; then
         echo "fail check text"
         exit 1
     fi
-    curl -q -m 6 -4 -s -L -H "X-header: $header" -H "X-magic: $magic" $urlt -o o.txt
-    curl -q -m 6 -4 -s -L -x socks5h://localhost:$lport -H "X-magic: $magic" $urlt?mask={1,12,123,1234} https://$urlt?mask={1,12,123,1234} -o - > o.magic.ws.txt
-    curl -q -m 6 -4 -s -L -x socks5h://localhost:$lport -H "X-header: $header" -H "X-magic: $magic" $urlt -o o.ws.txt
+    curl -q -m 6 -4 -s -L -x socks5h://$laddr:$lport -H "X-magic: $magic" $urlt?mask={1,12,123,1234} https://$urlt?mask={1,12,123,1234} -o - | grep '^X-magic:' > o.magic.ws.txt
+    curl -q -m 6 -4 -s -L --header @o.headers.txt -H "X-magic: $magic" $urlt -o o.txt
+    curl -q -m 6 -4 -s -L -x socks5h://$laddr:$lport --header @o.headers.txt -H "X-magic: $magic" $urlt -o o.ws.txt
     if ! cmp -s o.magic.txt o.magic.ws.txt; then
         echo "fail test text magic"
         exit 1
     fi
-    if ! cmp -s o.txt o.ws.txt; then
+    cat o.txt | grep -E '^(X-magic:|X-header-)' > o.headers.txt
+    cat o.ws.txt | grep -E '^(X-magic:|X-header-)' > o.headers.ws.txt
+    if ! cmp -s o.headers.txt o.headers.ws.txt; then
         echo "fail test text"
         exit 1
     fi
@@ -62,7 +64,9 @@ ss-local -l $lport -s 127.0.0.1 -p $sport -m chacha20-ietf-poly1305 -k sip003 --
 lpid=$!
 ss-server -s 127.0.0.1 -p $sport -m chacha20-ietf-poly1305 -k sip003 --plugin $v2ray_plugin --plugin-opts "server;mux=0" &
 spid=$!
-check
+if ! check; then
+    exit 1
+fi
 
 kill $spid
 echo -n wss-proxy client - wss-proxy server
@@ -73,7 +77,9 @@ done
 echo
 ss-server -s 127.0.0.1 -p $sport -m chacha20-ietf-poly1305 -k sip003 --plugin ./wss-proxy-server --plugin-opts "mux=0" &
 spid=$!
-check
+if ! check; then
+    exit 1
+fi
 
 kill $lpid
 echo v2ray-plugin client - wss-proxy server
@@ -84,4 +90,6 @@ done
 echo
 ss-local -l $lport -s 127.0.0.1 -p $sport -m chacha20-ietf-poly1305 -k sip003 --plugin $v2ray_plugin --plugin-opts "mux=0" &
 lpid=$!
-check
+if ! check; then
+    exit 1
+fi
